add (untested) wasm support
right now, the globals are pointers to function pointers,
which i don't think are callable from outside the wasm module
in the host language

+++ include/yabo/vtable.h
@@ -29,7 +29,7 @@ enum ReturnStatus {
 
 struct VTableHeader {
   int64_t head;
-  size_t deref_level;
+  int64_t deref_level;
   int64_t (*typecast_impl)(void *, const void *, uint64_t);
   size_t (*mask_impl)(void *);
   size_t size;
